Reorganize yolov4 files
