+++ models/demos/yolov4/runner/performant_runner_infra.py
@@ -8,8 +8,8 @@ from loguru import logger
 import ttnn
 from models.demos.yolov4.common import get_model_result, load_torch_model
 from models.demos.yolov4.post_processing import gen_yolov4_boxes_confs, get_region_boxes
-from models.demos.yolov4.ttnn.model_preprocessing import create_yolov4_model_parameters
-from models.demos.yolov4.ttnn.yolov4 import TtYOLOv4
+from models.demos.yolov4.tt.model_preprocessing import create_yolov4_model_parameters
+from models.demos.yolov4.tt.yolov4 import TtYOLOv4
 from models.utility_functions import divup, is_wormhole_b0
 from tests.ttnn.utils_for_testing import assert_with_pcc
