+++ models/demos/yolov4/tests/pcc/test_ttnn_downsample2.py
@@ -10,14 +10,13 @@ from loguru import logger
 
 import ttnn
 from models.demos.yolov4.common import load_torch_model
-from models.demos.yolov4.ttnn.downsample4 import Down4
-from models.demos.yolov4.ttnn.model_preprocessing import create_ds4_model_parameters
+from models.demos.yolov4.tt.downsample2 import Down2
+from models.demos.yolov4.tt.model_preprocessing import create_ds2_model_parameters
 from models.utility_functions import skip_for_grayskull
 from tests.ttnn.utils_for_testing import assert_with_pcc
 
 
 @skip_for_grayskull()
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 16384}], indirect=True)
 @pytest.mark.parametrize(
     "resolution",
     [
@@ -25,23 +24,25 @@ from tests.ttnn.utils_for_testing import assert_with_pcc
         (640, 640),
     ],
 )
-def test_down4(device, reset_seeds, model_location_generator, resolution):
+@pytest.mark.parametrize("device_params", [{"l1_small_size": 16384}], indirect=True)
+def test_down2(device, reset_seeds, model_location_generator, resolution):
     torch.manual_seed(0)
 
-    torch_model = load_torch_model(model_location_generator, module="down4")
+    torch_model = load_torch_model(model_location_generator, module="down2")
     if resolution == (320, 320):
-        torch_input = torch.randn((1, 256, 40, 40), dtype=torch.float)
+        torch_input = torch.randn((1, 64, 160, 160), dtype=torch.float)
     elif resolution == (640, 640):
-        torch_input = torch.randn((1, 256, 80, 80), dtype=torch.float)
+        torch_input = torch.randn((1, 64, 320, 320), dtype=torch.float)
     else:
-        raise ValueError(f"Unsupported resolution: {resolution}")
+        raise ValueError(f"Unsupported resolution {resolution}")
     ref = torch_model(torch_input)
 
-    parameters = create_ds4_model_parameters(torch_model, torch_input, resolution, device)
+    parameters = create_ds2_model_parameters(torch_model, torch_input, resolution, device)
+
+    ttnn_model = Down2(device, parameters, parameters.conv_args)
+
     torch_input = torch_input.permute(0, 2, 3, 1)
     ttnn_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16)
-
-    ttnn_model = Down4(device, parameters, parameters.conv_args)
 
     result_ttnn = ttnn_model(ttnn_input)
 
@@ -53,4 +54,4 @@ def test_down4(device, reset_seeds, model_location_generator, resolution):
     result = ttnn.to_torch(result_ttnn)
     result = result.permute(0, 3, 1, 2)
     result = result.reshape(ref.shape)
-    assert_with_pcc(result, ref, 0.90)  # PCC 0.90 - The PCC will improve once #3612 is resolved.
+    assert_with_pcc(result, ref, 0.99)
