+++ models/demos/yolov4/tests/pcc/test_ttnn_downsample4.py
@@ -10,13 +10,14 @@ from loguru import logger
 
 import ttnn
 from models.demos.yolov4.common import load_torch_model
-from models.demos.yolov4.ttnn.downsample5 import Down5
-from models.demos.yolov4.ttnn.model_preprocessing import create_ds5_model_parameters
+from models.demos.yolov4.tt.downsample4 import Down4
+from models.demos.yolov4.tt.model_preprocessing import create_ds4_model_parameters
 from models.utility_functions import skip_for_grayskull
 from tests.ttnn.utils_for_testing import assert_with_pcc
 
 
 @skip_for_grayskull()
+@pytest.mark.parametrize("device_params", [{"l1_small_size": 16384}], indirect=True)
 @pytest.mark.parametrize(
     "resolution",
     [
@@ -24,24 +25,24 @@ from tests.ttnn.utils_for_testing import assert_with_pcc
         (640, 640),
     ],
 )
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 16384}], indirect=True)
-def test_down5(device, reset_seeds, model_location_generator, resolution):
+def test_down4(device, reset_seeds, model_location_generator, resolution):
     torch.manual_seed(0)
 
-    torch_model = load_torch_model(model_location_generator, module="down5")
+    torch_model = load_torch_model(model_location_generator, module="down4")
     if resolution == (320, 320):
-        torch_input = torch.randn((1, 512, 20, 20), dtype=torch.float)
+        torch_input = torch.randn((1, 256, 40, 40), dtype=torch.float)
     elif resolution == (640, 640):
-        torch_input = torch.randn((1, 512, 40, 40), dtype=torch.float)
+        torch_input = torch.randn((1, 256, 80, 80), dtype=torch.float)
     else:
         raise ValueError(f"Unsupported resolution: {resolution}")
     ref = torch_model(torch_input)
 
-    parameters = create_ds5_model_parameters(torch_model, torch_input, resolution, device)
-
+    parameters = create_ds4_model_parameters(torch_model, torch_input, resolution, device)
     torch_input = torch_input.permute(0, 2, 3, 1)
     ttnn_input = ttnn.from_torch(torch_input, dtype=ttnn.bfloat16)
-    ttnn_model = Down5(device, parameters, parameters.conv_args)
+
+    ttnn_model = Down4(device, parameters, parameters.conv_args)
+
     result_ttnn = ttnn_model(ttnn_input)
 
     start_time = time.time()
@@ -52,4 +53,4 @@ def test_down5(device, reset_seeds, model_location_generator, resolution):
     result = ttnn.to_torch(result_ttnn)
     result = result.permute(0, 3, 1, 2)
     result = result.reshape(ref.shape)
-    assert_with_pcc(result, ref, 0.89)  # PCC 0.89 - The PCC will improve once #3612 is resolved.
+    assert_with_pcc(result, ref, 0.90)  # PCC 0.90 - The PCC will improve once #3612 is resolved.
