+++ models/demos/yolov4/tests/pcc/test_ttnn_head.py
@@ -10,8 +10,8 @@ from loguru import logger
 
 import ttnn
 from models.demos.yolov4.common import load_torch_model
-from models.demos.yolov4.ttnn.head import TtHead
-from models.demos.yolov4.ttnn.model_preprocessing import create_head_model_parameters
+from models.demos.yolov4.tt.head import TtHead
+from models.demos.yolov4.tt.model_preprocessing import create_head_model_parameters
 from models.utility_functions import skip_for_grayskull
 from tests.ttnn.utils_for_testing import assert_with_pcc
