+++ models/demos/yolov4/tests/pcc/test_ttnn_post_processing.py
@@ -7,7 +7,7 @@ import torch
 
 import ttnn
 from models.demos.yolov4.post_processing import gen_yolov4_boxes_confs
-from models.demos.yolov4.ttnn.genboxes import TtGenBoxes
+from models.demos.yolov4.tt.genboxes import TtGenBoxes
 from models.utility_functions import skip_for_grayskull
 from tests.ttnn.utils_for_testing import assert_with_pcc
