+++ models/demos/yolov4/tt/downsample1.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.demos.yolov4.ttnn.common import Conv
+from models.demos.yolov4.tt.common import Conv
 
 
 def sharded_concat(input_tensors, num_cores=64, dim=3):  # expected input tensors to be in fp16, RM, same (h*w)
