+++ models/demos/yolov4/tt/neck.py
@@ -2,10 +2,11 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import ttnn
-from models.demos.yolov4.ttnn.common import Conv
 import math
 from models.utility_functions import is_blackhole
+
+import ttnn
+from models.demos.yolov4.tt.common import Conv
 
 
 def determine_num_cores_for_upsample(nhw: int, width: int, max_cores=64) -> int:
