+++ tech_reports/YoloV4-TTNN/yolov4.md
@@ -1690,7 +1690,7 @@ fi
 # Google Drive file ID
 FILE_ID="1wv_LiFeCRYwtpkqREPeI13-gPELBDwuJ"
 # Output filename
-OUTPUT="tests/ttnn/integration_tests/yolov4/yolov4.pth"
+OUTPUT="models/demos/yolov4/tests/pcc/yolov4.pth"
 
 # Download the file
 gdown "https://drive.google.com/uc?id=${FILE_ID}" -O "${OUTPUT}"
