+++ tests/nightly/single_card/yolov4/test_e2e_performant.py
@@ -1,0 +1,1 @@
+../../../../models/demos/yolov4/tests/perf/test_e2e_performant.py
