--- tests/nightly/single_card/yolov4/test_yolov4_performant.py
@@ -1,1 +1,0 @@
-../../../../models/demos/wormhole/yolov4/test_yolov4_performant.py
